[BLD-9056]-moving chatcontroller and other chat functionality to the activity + fragments passes events to the activity with activity<->fragment interface
